Stop using symlinks with bootstrap-custom.css
Before this commit, we used `less` to build a bootstrap-custom.css target,
and placed that into build/. The src/bootstrap-custom/bootstrap-custom.css
was a symlink to that build/ version. The logic was that everything within
src/ should be tracked by git.

This patch teaches the .gitignore to ignore that file, and we just build
it directly into src/bootstrap-custom/. The reason is that yarn packaging
scripts, for example the scripts that we use to build an Electron version,
forcibly delete the contents of build/ on invocation. So that is not a good
place for us to stash our own build artifacts.

When we upgrade to Bootstrap4, we can stop using less, and then we will
no longer need to manually create this file outside of the normal build.