Auto-infer Age from BirthDate, but just use it as a hint on the Weigh-ins page.
At the moment, my thought is to just use this as a hint to the lifter and score
table to confirm both the MeetDate and BirthDate. If you don't re-type the hint,
it isn't used anywhere.

That isn't good enough for a release, but it's OK for the moment.

We should also find a way to get this on the Entry.prototype.