Create distinct state for attempt unit and bw unit
Just a first pass here, still need to regression test things and make
sure that state was separated properly. We'll also want to add a
translation for the new strings, particularly for spanish since the
requestor speaks spanish.  Since the string is very similar, we should
be able to reuse it with slight modifications.